requirements.txt
google-analytics-data==0.11.1
python-dotenv==0.20.0
tweepy==4.8.0